Turn on continuum-boot
Tmux Continuum takes care of setting a systemd unit for starting tmux
and maintaining it; all I have to do is turn it on. Now tmux will run on
startup.

In the past, I've shied away from doing this because of locale issues. A
simple DefaultEnvironment line in my systemd user.conf fixed it, so now
I can enjoy the benefits.