Re-enable continuum-boot
Now that the rest of my system has been properly configured to set up
the needed env for a tmux systemd user service, I've opted back to
enabling continuum-boot to start tmux after logging in.

In line with this, I now explicitly set mode-keys to vi, since
previously it defaulted to vi from being started in a shell with
VISUAL=vim.